Tests/LibWeb/Text/expected/WebAnimations/misc/multiple-animations.txt

    element 1 is animating
element 2 is animating
